+++ plugins/data-management/src/AddConnectionWidget/components/ConnectionTypeSelect.tsx
@@ -1,71 +1,62 @@
 import React, { useEffect } from 'react'
 import { IconButton, MenuItem, TextField } from '@mui/material'
-import OpenInNewIcon from '@mui/icons-material/OpenInNew'
+import { ConnectionType } from '@jbrowse/core/pluggableElementTypes'
 
-import ConnectionType from '@jbrowse/core/pluggableElementTypes/ConnectionType'
+// icons
+import OpenInNewIcon from '@mui/icons-material/OpenInNew'
 
 function ConnectionTypeSelect(props: {
   connectionTypeChoices: ConnectionType[]
-  connectionType: ConnectionType
-  setConnectionType: Function
+  connectionType?: ConnectionType
+  setConnectionType: (c?: ConnectionType) => void
 }) {
   const { connectionTypeChoices, connectionType, setConnectionType } = props
 
   useEffect(() => {
-    if (!connectionType.name) {
+    if (!connectionType) {
       setConnectionType(connectionTypeChoices[0])
     }
   })
 
-  function handleChange(
-    event: React.ChangeEvent<{ name?: string; value: unknown }>,
-  ): void {
-    setConnectionType(
-      connectionTypeChoices.find(
-        (connectionTypeChoice: ConnectionType) =>
-          connectionTypeChoice.name === event.target.value,
-      ),
-    )
-  }
-  if (!connectionType.name) {
-    return null
-  }
   return (
     <form autoComplete="off">
-      <TextField
-        value={connectionType.name}
-        label="connectionType"
-        helperText={
-          connectionType.description ? (
-            <>
-              {connectionType.description}
-              {connectionType.url ? (
-                <IconButton
-                  href={connectionType.url}
-                  rel="noopener noreferrer"
-                  target="_blank"
-                  color="secondary"
-                >
-                  <OpenInNewIcon />
-                </IconButton>
-              ) : null}
-            </>
-          ) : null
-        }
-        select
-        fullWidth
-        onChange={handleChange}
-        variant="outlined"
-      >
-        {connectionTypeChoices.map((connectionTypeChoice: ConnectionType) => (
-          <MenuItem
-            key={connectionTypeChoice.name}
-            value={connectionTypeChoice.name}
-          >
-            {connectionTypeChoice.displayName || connectionTypeChoice.name}
-          </MenuItem>
-        ))}
-      </TextField>
+      {connectionType ? (
+        <TextField
+          value={connectionType.name}
+          label="connectionType"
+          helperText={
+            connectionType.description ? (
+              <>
+                {connectionType.description}
+                {connectionType.url ? (
+                  <IconButton
+                    href={connectionType.url}
+                    rel="noopener noreferrer"
+                    target="_blank"
+                    color="secondary"
+                  >
+                    <OpenInNewIcon />
+                  </IconButton>
+                ) : null}
+              </>
+            ) : null
+          }
+          select
+          fullWidth
+          onChange={event =>
+            setConnectionType(
+              connectionTypeChoices.find(c => c.name === event.target.value),
+            )
+          }
+          variant="outlined"
+        >
+          {connectionTypeChoices.map(c => (
+            <MenuItem key={c.name} value={c.name}>
+              {c.displayName || c.name}
+            </MenuItem>
+          ))}
+        </TextField>
+      ) : null}
     </form>
   )
 }
