+++ plugins/data-management/src/AddConnectionWidget/model.ts
